tests/wpt/tests/css/css-anchor-position/anchor-position-multicol-002.html

<!DOCTYPE html>
<link rel="help" href="https://drafts.csswg.org/css-anchor-1/#propdef-anchor-name">
<link rel="help" href="https://drafts.csswg.org/css-anchor-1/#anchor-pos">
<link rel="help" href="https://drafts.csswg.org/css-anchor-1/#anchor-size">
<link rel="author" href="mailto:kojii@chromium.org">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/check-layout-th.js"></script>
<script src="support/test-common.js"></script>
<style>
.cb {
  transform: translate(0);  /* Make it a containing block. */
  border-width: 5px 6px 7px 8px;
  border-style: solid;
  padding: 5px 6px 7px 8px;
}
.columns {
  column-count: 4;
  column-fill: auto;
  column-gap: 10px;
  column-width: 100px;
  width: 320px;
  height: 100px;
}
.spacer {
  height: 10px;
  background: silver;
}
.anchor1 {
  anchor-name: --a1;
  margin-left: 10px;
  width: 50px;
  height: 90px;
  background: yellow;
}
.target {
  position: absolute;
}
.fixed {
  position: fixed;
}
.target1 {
  left: anchor(--a1 left);
  top: anchor(--a1 top);
  width: anchor-size(--a1 width);
  height: anchor-size(--a1 height);
  border: dashed 4px aqua;
  opacity: .2;
  box-sizing: border-box;
}
.target1-rb {
  right: anchor(--a1 right);
  bottom: anchor(--a1 bottom);
  width: 10px;
  height: 10px;
  background: teal;
}

.inner .target1 {
  border-color: fuchsia;
  border-style: solid;
}
.inner .target1-rb {
  background: purple;
}

</style>
<body onload="checkLayoutForAnchorPos('.target')">
  <p>The dashed aqua rectangle should encompass both yellow fragments,
    and the teal box should be at the bottom-right of the aqua box.
  <p>The fuchsia rectangle should match the edge of the yellow box except where it's split across columns,
    and the purple box should be at the bottom-right of the fuchsia rectangle.
  <div class="spacer" style="height: 10px"></div>
  <div class="cb outer">
    <div class="columns">
      <div class="spacer" style="height: 30px"></div>
      <div class="cb inner">
        <!-- This spacer fills up to the middle of the 2nd column. -->
        <div class="spacer" style="height: 130px"></div>
        <div class="anchor1"></div>
        <div class="spacer" style="height: 100px"></div>

        <!-- The containing block of querying elements is block-fragmented. -->
        <div class="target target1"
             data-offset-x=18 data-offset-y=135
             data-expected-width=50 data-expected-height=90></div>
        <div class="target target1-rb"
             data-offset-x=58 data-offset-y=215></div>
        <div class="target fixed target1"
             data-offset-x=26 data-offset-y=140
             data-expected-width=50 data-expected-height=90></div>
        <div class="target fixed target1-rb"
             data-offset-x=58 data-offset-y=215></div>
      </div>

      <!-- The containing block of querying elements is a multi-column.  -->
      <div class="target target1"
           data-offset-x=144 data-offset-y=5
           data-expected-width=160 data-expected-height=100></div>
      <div class="target target1-rb"
           data-offset-x=294 data-offset-y=95></div>
      <div class="target fixed target1"
           data-offset-x=144 data-offset-y=5
           data-expected-width=160 data-expected-height=100></div>
      <div class="target fixed target1-rb"
           data-offset-x=294 data-offset-y=95></div>
    </div>

    <!-- The containing block of querying elements is not fragmented.  -->
    <div class="target target1"
         data-offset-x=144 data-offset-y=5
         data-expected-width=160 data-expected-height=100></div>
    <div class="target target1-rb"
         data-offset-x=294 data-offset-y=95></div>
    <div class="target fixed target1"
         data-offset-x=144 data-offset-y=5
         data-expected-width=160 data-expected-height=100></div>
    <div class="target fixed target1-rb"
         data-offset-x=294 data-offset-y=95></div>
  </div>
</body>
